pyproject.toml
[tool.poetry]
name = "crm"
version = "0.1.0"
description = ""
authors = ["Magomed Usmanov <mgsmnov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
djangorestframework = "^3.15.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
